Rough draft of Remote Config public API
This serves as a generic interface that we can use throughout the application, even before an implementation of the remote config source exists.

As a rough draft, some possible issues are expected until we iterate on the design.

The test can be run from the command line by navigating to the ConfigurationKit directory, then running `swift test`